garment_core/erdeniz_security/requirements-windows.txt
--only-binary cryptography
--only-binary argon2-cffi
cryptography>=42.0.0
argon2-cffi>=23.1.0
bcrypt>=4.1.0
django>=4.2
django-encrypted-model-fields>=0.6.5
python-decouple>=3.8
